Renamed dateFormatHelpers -> dateHelpers
