+++ src/components/sidebar/lineView/LineHeaderTable.tsx
@@ -12,7 +12,7 @@ import { AlertStore } from '~/stores/alertStore';
 import { ConfirmStore } from '~/stores/confirmStore';
 import { ErrorStore } from '~/stores/errorStore';
 import { IMassEditLineHeader, LineHeaderMassEditStore } from '~/stores/lineHeaderMassEditStore';
-import { getDateWithoutHours } from '~/util/dateFormatHelpers';
+import { getDateWithoutHours } from '~/util/dateHelpers';
 import FormValidator from '~/validation/FormValidator';
 import SidebarHeader from '../SidebarHeader';
 import LineHeaderForm from './LineHeaderForm';
